+++ satellite/src/main/java/com/linbit/linstor/core/objects/ExternalFileSatelliteFactory.java
@@ -17,6 +17,7 @@ import com.linbit.linstor.utils.ByteUtils;
 import javax.inject.Inject;
 import javax.inject.Provider;
 
+import java.util.List;
 import java.util.UUID;
 
 public class ExternalFileSatelliteFactory
@@ -48,7 +49,8 @@ public class ExternalFileSatelliteFactory
         UUID uuid,
         ExternalFileName extFileNameRef,
         long initflags,
-        @Nullable byte[] content
+        @Nullable byte[] content,
+        List<String> altSuffixesRef
     )
         throws ImplementationError
     {
@@ -64,6 +66,7 @@ public class ExternalFileSatelliteFactory
                     initflags,
                     content == null ? new byte[0] : content,
                     content == null ? new byte[0] : ByteUtils.checksumSha256(content),
+                    altSuffixesRef,
                     driver,
                     transObjFactory,
                     transMgrProvider