+++ server/src/main/java/com/linbit/linstor/dbdrivers/interfaces/ExternalFileDatabaseDriver.java
@@ -1,6 +1,7 @@
 package com.linbit.linstor.dbdrivers.interfaces;
 
 import com.linbit.linstor.core.objects.ExternalFile;
+import com.linbit.linstor.dbdrivers.interfaces.updater.CollectionDatabaseDriver;
 import com.linbit.linstor.dbdrivers.interfaces.updater.SingleColumnDatabaseDriver;
 import com.linbit.linstor.stateflags.StateFlagsPersistence;
 
@@ -11,4 +12,6 @@ public interface ExternalFileDatabaseDriver extends GenericDatabaseDriver<Extern
     SingleColumnDatabaseDriver<ExternalFile, byte[]> getContentCheckSumDriver();
 
     StateFlagsPersistence<ExternalFile> getStateFlagPersistence();
+
+    CollectionDatabaseDriver<ExternalFile, String> getAltSuffixesDriver();
 }